+++ internal/librarian/php/generate_test.go
@@ -1,0 +1,33 @@
+// Copyright 2026 Google LLC
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//     https://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+package php
+
+import (
+	"testing"
+
+	"github.com/googleapis/librarian/internal/config"
+	"github.com/googleapis/librarian/internal/sources"
+)
+
+func TestGenerate(t *testing.T) {
+	ctx := t.Context()
+	cfg := &config.Config{}
+	lib := &config.Library{}
+	src := &sources.Sources{}
+
+	if err := Generate(ctx, cfg, lib, src); err != nil {
+		t.Errorf("Generate() returned error: %v", err)
+	}
+}
